reinstate treeshrink lineages
On the 14/6/20 tree, a few of us noticed that lineage A.3 had been misplaced. It looks a lot like some previously-identified long-branch taxa are partly responsible for this. So I decided to reinstate the iterative removal of treeshrink taxa, while simultaneously re-anslysing the 14/6/20 datatset to see if this can be avoided in future via changing the analysis settings.